refactor: change annotations names and folders
Took 4 hours 45 minutes
